mqtt/requirements.txt
# ESP32 MQTT Client Python Dependencies
# Install with: pip install -r requirements.txt

paho-mqtt>=1.6.0